pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyrexia-stat"
version = "0.9.4"
authors = [
    { name="Edward McGrath", email="edward@edwardmcgrath.com" },
]
description = "Thermostat runs on raspberry pi"
readme = "README.md"
dependencies = [
    "bleak>=0.19.5",
    "pytz>=2022.2.1",
    "sensorpush-bleak>=1.0.7",
    "RPi.GPIO>=0.7.0",
    "Adafruit-DHT>=1.4.0",
    "requests>=2.25.1",

]
requires-python = ">3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/cryptomcgrath/pyrexia-stat"

